rename scalar
