+++ source/consumers/dataspaces/CMakeLists.txt
@@ -3,7 +3,7 @@ set(dir ${CMAKE_CURRENT_SOURCE_DIR})
 set(SOURCES
   ${dir}/dense1d.cpp
   ${dir}/dense_matrix2d.cpp
-  ${dir}/scalar.cpp
+  ${dir}/Scalar.cpp
   ${dir}/SparseMap2D.cpp
   ${dir}/SparseMap3D.cpp
   ${dir}/SparseMatrix2D.cpp
@@ -12,7 +12,7 @@ set(SOURCES
 set(HEADERS
   ${dir}/dense1d.h
   ${dir}/dense_matrix2d.h
-  ${dir}/scalar.h
+  ${dir}/Scalar.h
   ${dir}/SparseMap2D.h
   ${dir}/SparseMap3D.h
   ${dir}/SparseMatrix2D.h
