+++ tests/consumers/dataspaces/CMakeLists.txt
@@ -2,7 +2,7 @@ set(dir ${CMAKE_CURRENT_SOURCE_DIR})
 
 set(SOURCES
   ${dir}/dense1d.cpp
-  ${dir}/scalar.cpp
+  ${dir}/ScalarTest.cpp
   ${dir}/SparseMap2DTest.cpp
   ${dir}/SparseMap3DTest.cpp
   ${dir}/SparseMatrix2DTest.cpp
